JVM: remove K1 usages from serialization.jvm
Unfortunately, JvmDescriptorMangler was used both from
serialization.jklib and scripting-compiler, so I had to copy it to both,
but the latter usage is destined to be removed in KT-87149.

 #KT-84406
 #KT-84562
