+++ compiler/ir/serialization.jvm/build.gradle.kts
@@ -8,11 +8,9 @@ plugins {
 dependencies {
     api(project(":compiler:ir.tree"))
     api(project(":compiler:ir.serialization.common"))
-    implementation(project(":compiler:frontend"))
-    implementation(project(":core:descriptors"))
-    implementation(project(":core:descriptors.jvm"))
+    implementation(project(":compiler:config.jvm"))
+    implementation(project(":core:compiler.common.jvm"))
     implementation(project(":kotlin-util-klib-metadata"))
-    implementation(project(":compiler:frontend.java"))
 }
 
 optInToUnsafeDuringIrConstructionAPI()
